+++ src/stock-lookups/stock-lookups.resource.ts
@@ -25,18 +25,16 @@ export type ConceptFilterCriteria = ResourceFilterCriteria;
 export type LocationFilterCriteria = ResourceFilterCriteria;
 
 // getLocations
-export function useStockLocations(filter: LocationFilterCriteria) {
+export function useLocations(filter: LocationFilterCriteria) {
   const apiUrl = `ws/rest/v1/location${toQueryParams(filter, false)}`;
   const { data, error, isLoading } = useSWR<
-    {
-      data: PageableResult<OpenMRSLocation>;
-    },
+    { data: PageableResult<OpenMRSLocation> },
     Error
   >(apiUrl, openmrsFetch);
   return {
-    locations: data?.data || <PageableResult<OpenMRSLocation>>{},
-    isErrorLocation: error,
-    isLoadingLocations: isLoading,
+    items: data?.data || <PageableResult<OpenMRSLocation>>{},
+    isLoading,
+    isError: error,
   };
 }
 
@@ -44,9 +42,7 @@ export function useStockLocations(filter: LocationFilterCriteria) {
 export function useLocationWithIdByUuid(id: string) {
   const apiUrl = `ws/rest/v1/stockmanagement/location/${id}`;
   const { data, error, isLoading } = useSWR<
-    {
-      data: PageableResult<OpenMRSLocation>;
-    },
+    { data: PageableResult<OpenMRSLocation> },
     Error
   >(apiUrl, openmrsFetch);
   return {
@@ -75,9 +71,7 @@ export function useLocationTags(q: string) {
     q && q.length > 0 ? "&q=" + encodeURIComponent(q) : ""
   }`;
   const { data, error, isLoading } = useSWR<
-    {
-      data: PageableResult<OpenMRSLocationTag>;
-    },
+    { data: PageableResult<OpenMRSLocationTag> },
     Error
   >(apiUrl, openmrsFetch);
   return {
@@ -91,13 +85,11 @@ export function useLocationTags(q: string) {
 export function useRoles(filter: ResourceFilterCriteria) {
   const apiUrl = `ws/rest/v1/role${toQueryParams(filter)}`;
   const { data, error, isLoading } = useSWR<
-    {
-      data: PageableResult<Role>;
-    },
+    { data: PageableResult<Role> },
     Error
   >(apiUrl, openmrsFetch);
   return {
-    items: data?.data || <PageableResult<Role>>{},
+    items: data.data ? data.data : [],
     isLoading,
     isError: error,
   };
@@ -106,14 +98,12 @@ export function useRoles(filter: ResourceFilterCriteria) {
 // getStockOperationTypes
 export function useStockOperationTypes() {
   const apiUrl = `ws/rest/v1/stockmanagement/stockoperationtype?v=default`;
-  const { data, isLoading, error } = useSWR<
-    {
-      data: PageableResult<StockOperationType>;
-    },
+  const { data, error, isLoading } = useSWR<
+    { data: PageableResult<StockOperationType> },
     Error
   >(apiUrl, openmrsFetch);
   return {
-    types: data?.data || <PageableResult<StockOperationType>>{},
+    items: data.data ? data.data : [],
     isLoading,
     isError: error,
   };
@@ -123,13 +113,11 @@ export function useStockOperationTypes() {
 export function useUsers(filter: UserFilterCriteria) {
   const apiUrl = `ws/rest/v1/user${toQueryParams(filter)}`;
   const { data, error, isLoading } = useSWR<
-    {
-      data: PageableResult<User>;
-    },
+    { data: PageableResult<User> },
     Error
   >(apiUrl, openmrsFetch);
   return {
-    items: data?.data || <PageableResult<User>>{},
+    items: data.data ? data.data : [],
     isLoading,
     isError: error,
   };
@@ -138,12 +126,10 @@ export function useUsers(filter: UserFilterCriteria) {
 // getUser
 export function useUser(id: string) {
   const apiUrl = `ws/rest/v1/user${id}`;
-  const { data, error, isLoading } = useSWR<
-    {
-      data: User;
-    },
-    Error
-  >(apiUrl, openmrsFetch);
+  const { data, error, isLoading } = useSWR<{ data: User }, Error>(
+    apiUrl,
+    openmrsFetch
+  );
   return {
     items: data.data ? data.data : {},
     isLoading,
@@ -154,12 +140,10 @@ export function useUser(id: string) {
 // getConceptById
 export function useConceptById(id: string) {
   const apiUrl = `ws/rest/v1/concept/${id}`;
-  const { data, error, isLoading } = useSWR<
-    {
-      data: Concept;
-    },
-    Error
-  >(apiUrl, openmrsFetch);
+  const { data, error, isLoading } = useSWR<{ data: Concept }, Error>(
+    apiUrl,
+    openmrsFetch
+  );
   return {
     items: data?.data || <Concept>{},
     isLoading,
@@ -171,9 +155,7 @@ export function useConceptById(id: string) {
 export function useParties() {
   const apiUrl = `ws/rest/v1/stockmanagement/party?v=default`;
   const { data, error, isLoading } = useSWR<
-    {
-      data: PageableResult<Party>;
-    },
+    { data: PageableResult<Party> },
     Error
   >(apiUrl, openmrsFetch);
   return {
@@ -187,13 +169,11 @@ export function useParties() {
 export function useDrugs(filter: DrugFilterCriteria) {
   const apiUrl = `ws/rest/v1/drug${toQueryParams(filter)}`;
   const { data, error, isLoading } = useSWR<
-    {
-      data: PageableResult<Drug>;
-    },
+    { data: PageableResult<Drug> },
     Error
   >(apiUrl, openmrsFetch);
   return {
-    items: data?.data || <PageableResult<Drug>>{},
+    items: data.data ? data.data : [],
     isLoading,
     isError: error,
   };
@@ -203,13 +183,11 @@ export function useDrugs(filter: DrugFilterCriteria) {
 export function useConcepts(filter: ConceptFilterCriteria) {
   const apiUrl = `ws/rest/v1/concept${toQueryParams(filter)}`;
   const { data, error, isLoading } = useSWR<
-    {
-      data: PageableResult<Concept>;
-    },
+    { data: PageableResult<Concept> },
     Error
   >(apiUrl, openmrsFetch);
   return {
-    items: data?.data || <PageableResult<Concept>>{},
+    items: data.data ? data.data : [],
     isLoading,
     isError: error,
   };
@@ -219,9 +197,7 @@ export function useConcepts(filter: ConceptFilterCriteria) {
 export function usePatients(filter: ConceptFilterCriteria) {
   const apiUrl = `ws/rest/v1/patient${toQueryParams(filter)}`;
   const { data, error, isLoading } = useSWR<
-    {
-      data: PageableResult<Patient>;
-    },
+    { data: PageableResult<Patient> },
     Error
   >(apiUrl, openmrsFetch);
   return {
