+++ src/stock-operations/add-stock-operation/add-stock-operation.scss
@@ -1,6 +1,6 @@
-@use '@carbon/styles/scss/colors';
 @use '@carbon/styles/scss/spacing';
 @use '@carbon/styles/scss/type';
+@import '~@openmrs/esm-styleguide/src/vars';
 
 section {
   margin: spacing.$spacing-03;
@@ -8,7 +8,7 @@ section {
 
 .sectionTitle {
   @include type.type-style('heading-compact-02');
-  color: colors.$gray-70;
+  color: $text-02;
   margin-bottom: spacing.$spacing-04;
 }
