--- src/stock-sources/stock-sources-add-or-create-dialog/stock-sources-add-or-create-dialog.scss
@@ -1,17 +1,0 @@
-@use '@carbon/styles/scss/spacing';
-@use '@carbon/styles/scss/type';
-@import '~@openmrs/esm-styleguide/src/vars';
-
-section {
-  margin: spacing.$spacing-03;
-}
-
-.sectionTitle {
-  @include type.type-style('heading-compact-02');
-  color: $text-02;
-  margin-bottom: spacing.$spacing-04;
-}
-
-.modalBody {
-  padding-bottom: spacing.$spacing-05;
-}
