+++ src/stock-user-role-scopes/add-stock-user-scope/add-stock-user-role-scope.scss
@@ -1,13 +1,14 @@
-@use '@carbon/styles/scss/colors';
 @use '@carbon/styles/scss/spacing';
 @use '@carbon/styles/scss/type';
+@import '~@openmrs/esm-styleguide/src/vars';
 
 section {
   margin: spacing.$spacing-03;
 }
 
 .sectionTitle {
-  color: colors.$gray-70;
+  @include type.type-style('heading-compact-02');
+  color: $text-02;
   margin-bottom: spacing.$spacing-04;
 }
